Extract core print logic
This extracts the core print logic away from the React component
and makes it available as plain JS. This allows us the ability to
export the functionality so that end users can use it without
having to actually use it in a JSX context - they can now print via
the API.
